Add proptest for string list pool
