+++ lading_payload/src/common/strings/random_string_pool.rs
@@ -187,3 +187,59 @@ where
     }
     buf
 }
+
+#[cfg(test)]
+mod test {
+    use proptest::prelude::*;
+
+    use super::{ALPHANUM, RandomStringPool};
+    use rand::{SeedableRng, rngs::SmallRng};
+
+    // Ensure that no returned string ever has a non-alphabet character.
+    proptest! {
+        #[test]
+        fn no_nonalphabet_char(seed: u64, max_bytes: u16, of_size_bytes: u16) {
+            let max_bytes = max_bytes as usize;
+            let of_size_bytes = of_size_bytes as usize;
+            let mut rng = SmallRng::seed_from_u64(seed);
+
+            let pool = RandomStringPool::with_size_and_alphabet(&mut rng, max_bytes, ALPHANUM);
+            if let Some(s) = pool.of_size(&mut rng, of_size_bytes) {
+                for c in s.bytes() {
+                    assert!(ALPHANUM.contains(&c));
+                }
+            }
+        }
+    }
+
+    // Ensure that no returned string is ever larger or smaller than of_size_bytes.
+    proptest! {
+        #[test]
+        fn no_size_mismatch(seed: u64, max_bytes: u16, of_size_bytes: u16) {
+            let max_bytes = max_bytes as usize;
+            let of_size_bytes = of_size_bytes as usize;
+            let mut rng = SmallRng::seed_from_u64(seed);
+
+            let pool = RandomStringPool::with_size_and_alphabet(&mut rng, max_bytes, ALPHANUM);
+            if let Some(s) = pool.of_size(&mut rng, of_size_bytes) {
+                assert!(s.len() == of_size_bytes);
+            }
+        }
+    }
+
+    // Ensure that of_size only returns None if the request is greater than or
+    // equal to the interior size.
+    proptest! {
+        #[test]
+        fn return_none_condition(seed: u64, max_bytes: u16, of_size_bytes: u16) {
+            let max_bytes = max_bytes as usize;
+            let of_size_bytes = of_size_bytes as usize;
+            let mut rng = SmallRng::seed_from_u64(seed);
+
+            let pool = RandomStringPool::with_size_and_alphabet(&mut rng, max_bytes, ALPHANUM);
+            if pool.of_size(&mut rng, of_size_bytes).is_none() {
+                assert!(of_size_bytes >= max_bytes);
+            }
+        }
+    }
+}
